% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbUncertainty.R
\name{diffscoreciALT}
\alias{diffscoreciALT}
\title{Calculate score confidence intervals of the difference in probability.
Internal function of scoreCI, 2.}
\usage{
diffscoreciALT(x1, n1, x2, n2, conf.level)
}
\arguments{
\item{x1}{Success counts in sample 1}

\item{n1}{Sample size in sample 1}

\item{x2}{Success counts in sample 2}

\item{n2}{Sample size in sample 2}

\item{conf.level}{Confidence coefficient}
}
\description{
Calculate score confidence intervals of the difference in probability.
Internal function of scoreCI, 2.
}
\details{
Adapted from the internal 'z2stat' function of 'diffscoreci' from the R package
{PropCIs}. For extreme differences, the diffscoreci can throw up errors, which is
ultimately a result of a machine error in floating point math. More specifically,
the ratio \eqn{v/u^3} should never be above one, but it sometimes can be with
floating point error. This internal function should correct it.
}
\references{
Scherer, R. (2018). PropCIs: Various Confidence Interval Methods for Proportions. \url{https://cran.r-project.org/package=PropCIs}
}
