% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMuncertainty.R
\name{glm.prob.diff}
\alias{glm.prob.diff}
\title{calculate difference in probabilities between two groups from GLM output}
\usage{
glm.prob.diff(m)
}
\arguments{
\item{m}{is the GLM object}
}
\description{
calculate difference in probabilities between two groups from GLM output
}
