% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMuncertainty.R
\name{glm.se.data}
\alias{glm.se.data}
\title{Calculate the standard error of a difference between two properties
from a GLM analysis using the delta method}
\usage{
glm.se.data(m)
}
\arguments{
\item{m}{is the GLM object}
}
\description{
Calculate the standard error of a difference between two properties
from a GLM analysis using the delta method
}
\references{
Dorfman, R. (1938). A note on the delta-method for finding variance formulae. The Biométrie Bulletin, 1, 129–137

Ver Hoef, J. M. (2012). Who invented the delta method? American Statistician, 66(2), 124–127. https://doi.org/10.1080/00031305.2012.687494
}
