% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMuncertainty.R
\name{inv.logit}
\alias{inv.logit}
\title{translate GLM coef from the logit / log-odds scale to the probability scale}
\usage{
inv.logit(x)
}
\arguments{
\item{x}{coefficient in GLM object}
}
\description{
translate GLM coef from the logit / log-odds scale to the probability scale
}
