% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMuncertainty.R
\name{profCI}
\alias{profCI}
\title{Calculate the profile likelihood-based confidence intervals for a difference
in probabilities between two groups.To generate the profile likelihood CI, we
fixed δp to a range of values between -0.99 and 0.99, and found the value of
p0  that maximised the likelihood of observing the data, given the fixed
value of δp. We then recorded the upper and lower values of δ_p for which the
twice the log likelihood difference from the unconstrained model was less than
the 95\% quantile of a chi-squared distribution with one degree of freedom.}
\usage{
profCI(x, y)
}
\arguments{
\item{x}{is predictor variable}

\item{y}{is response variable}
}
\description{
Calculate the profile likelihood-based confidence intervals for a difference
in probabilities between two groups.To generate the profile likelihood CI, we
fixed δp to a range of values between -0.99 and 0.99, and found the value of
p0  that maximised the likelihood of observing the data, given the fixed
value of δp. We then recorded the upper and lower values of δ_p for which the
twice the log likelihood difference from the unconstrained model was less than
the 95\% quantile of a chi-squared distribution with one degree of freedom.
}
\references{
Cole, S. R., Chu, H., & Greenland, S. (2014). Maximum likelihood, profile likelihood, and penalized likelihood: A primer. American Journal of Epidemiology, 179(2), 252–260. https://doi.org/10.1093/aje/kwt245
}
