% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMuncertainty.R
\name{scoreCI}
\alias{scoreCI}
\title{Calculate the score confidence interval. In contrast to Wald-type CI, which
uses the estimated standard error to calculate CIs, the score method uses
the null standard error}
\usage{
scoreCI(x, y)
}
\arguments{
\item{x}{is predictor variable}

\item{y}{is response variable}
}
\description{
Calculate the score confidence interval. In contrast to Wald-type CI, which
uses the estimated standard error to calculate CIs, the score method uses
the null standard error
}
\references{
Wilson, E. B. . (1927). Probable Inference, the Law of Succession, and Statistical Inference. Journal of the American Statistical Association, 22(158), 209–212. https://doi.org/10.2307/2276774

Brown, L. D., Cai, T. T., & DasGupta, A. (2001). Interval estimation of binomial proportion. Statistical Science, 16(2), 101–133. https://doi.org/10.1214/ss/1009213286

Scherer, R. (2018). PropCIs: Various Confidence Interval Methods for Proportions. https://cran.r-project.org/package=PropCIs
}
