% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{startrek}
\alias{startrek}
\title{Star Trek The Original Series - Survival of Starfleet personnel}
\format{
A data frame with 430 rows and 2 variables:
\describe{
\item{uniform}{each row is a unique individual and the color of their Starfleet uniform (gold, blue, or red)}
\item{status}{survival status of each individual at the end of Season 3 (dead or alive)}
}
}
\source{
\url{https://www.ex-astris-scientia.org/database/redshirt_deaths.htm}

\url{https://memory-gamma.fandom.com/wiki/USS_Enterprise_(NCC-1701)}
}
\usage{
startrek
}
\description{
A dataset containing the survival of the 430 crew on-board the Enterprise,
including 43 officers (gold uniforms), 151 scientists (blue uniforms),
and 236 members of the operations division (red uniforms)
}
\examples{
data(startrek)
}
\keyword{datasets}
