% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMuncertainty.R
\name{z2statALT}
\alias{z2statALT}
\title{Calculate score confidence intervals of the difference in probability.
Internal function of scoreCI, 1.}
\usage{
z2statALT(p1x, nx, p1y, ny, dif)
}
\arguments{
\item{p1x}{Ratio of success counts in sample 1 / sample size in sample 1}

\item{nx}{Sample size in sample 1}

\item{p1y}{Ratio of success counts in sample 2 / sample size in sample 2}

\item{ny}{Sample size in sample 2}

\item{dif}{(p1x - p1y) - (0.5 * (1 + (p1x - p1y)))}
}
\description{
Calculate score confidence intervals of the difference in probability.
Internal function of scoreCI, 1.
}
\details{
Adapted from the internal 'z2stat' function of 'diffscoreci' from the R package
{PropCIs}. For extreme differences, the 'diffscoreci' can throw up errors, which is
ultimately a result of a machine error in floating point math. More specifically,
the ratio \eqn{v/u^3} should never be above one, but it sometimes can be with
floating point error. This internal function should correct the issue.
}
\references{
Scherer, R. (2018). PropCIs: Various Confidence Interval Methods for Proportions. \url{https://cran.r-project.org/package=PropCIs}
}
