% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMuncertainty.R
\name{diffscoreciALT}
\alias{diffscoreciALT}
\title{Calculate the score confidence interval. In contrast to Wald-type CI, which
uses the estimated standard error to calculate CIs, the score method uses
the null standard error. A wrapper.}
\usage{
diffscoreciALT(x1, n1, x2, n2, conf.level)
}
\arguments{
\item{x1}{defined in scoreCI function below}

\item{n1}{defined in scoreCI function below}

\item{x2}{defined in scoreCI function below}

\item{n2}{defined in scoreCI function below}

\item{conf.level}{defined in scoreCI function below}
}
\description{
Calculate the score confidence interval. In contrast to Wald-type CI, which
uses the estimated standard error to calculate CIs, the score method uses
the null standard error. A wrapper.
}
