% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbUncertainty.R
\name{inv.logit}
\alias{inv.logit}
\title{Inverse of the logit link function.}
\usage{
inv.logit(x)
}
\arguments{
\item{x}{A numeric object.}
}
\value{
A numeric object containing the inverse logit of the input value.
}
\description{
Inverse of the logit link function.
}
\details{
The inverse logit is defined by \eqn{exp(x)/(1+exp(x))}. Values in x are transformed
from a real number (the logarithm of the odds; value between -Inf and Inf) to a probability
with a value between 0 and 1.
}
